Fix calling issues with alignment and structs
Structs that were passed by value were previously not being assigned to
the stack correctly. This commit modifies the CallConv TableGen file to
assign them correctly. They have a minimum size of 4 and a minimum
alignment of 4.

Additionally, this commit tells LLVM that this backend doesn't support
stack realignment. It also expands the instructions used to implement
stack alignment without that present.