Make the frame pointer point to the "first local"
Previously, the first four bytes of the local area were completely
unused. Now, the frame pointer points four bytes above where the old
frame pointer is stored. Note that this only concides with the first local
variable if that stack slot is an integer. If it is a byte for example, it will
be at FP+3. Likewise if it is a short, it will be at FP+2.