Fix struct return redundancy
It is possible for us to have to sret even though the LLVM function is
not marked as sret. For example, if we want to return a double or an i64.
Previously, the code that stored the sret address into the return address
slot only checked the LLVM function's attributes. Now, it checks whether
we have any parameters marked for sret.