Add git credentials to build container
The build script checks out the repository in order to build. However,
that doesn't immediately work since our repository is private. To get
around this, we use BuildKit's secrets. We create a git credential
helper that echoes the username and password, and we pass that as a
secret into the build container.