Add scaffolding for TestElision pass
Additionally, add calls to `initializeXXXPass` wherever possible. It's
not entirely clear what these calls do, but it's probably good to have
them. They are idempotent, so we add calls both when initializing the
target and when constructing the pass.

This commit also implements the `getPassName` method on the AsmPrinter.
That's nice to have.