Add patterns to mask via shifts
If we want to zero out just the uppermost or the lowermost bits, it
might be profitable to use shift instructions to do that. This commit
adds patterns in TableGen to select them.

This might cause slowdowns in instruction selection. If this becomes an
issue, we should either select these patterns with custom C++, or we
should ditch them entirely.