Patch comparison libcall implementation
Previously, the libcalls could falsely report equality when zero was
supplied the LHS or RHS of the signed and unsigned comparisons
functions respectively. This would cause, for example `0` to be equal to
any negative number when signed. This has been patched.

Additionally, this commit implements a performance improvement for
unsigned comparison.