Always reserve a scavenging stack slot
Previously, we would dynamically decide whether the function needs an
emergency stack slot. However, this code was not able to account for all
the different cases. As such, we now pessimistically assume we'll always
need an emergency stack slot.

In the future, it might be nice to allow functions to decide that they
don't need the emergency stack slot via an attribute.