Fix issue with empty basic blocks causing crashes
Previously, generating a large immediate add would look at the parent of
the iterator passed in to determine the basic block being worked on.
This doesn't work if the iterator passed in is `end`, which it can be if
the first block in the function is empty. Thus, we now take in the basic
block explicitly.